Desenvolvimento - Menu de opções / Ajustes

+++ crud.py
@@ -69,31 +69,32 @@ def listar ():
     # Criando um loop para contabilizar as linhas do banco (row), utlizndo método - view
     if db:
         # Se não houver registros, então...
-        if db.info()['doc_count'] == 0:
+        if db.info()['doc_count'] < 1:
             print('\n>>> Não há usuários registrados <<<')
-        else:    
-            for row in db.view('all_docs', include_docs=True):
+        else:
+            print('\t-------------------------- REGISTROS ---------------------------------')    
+            for row in db.view('_all_docs', include_docs=True):
                 # Atribui-se ao documento (doc) a lista obtida
                 doc = row['doc']
                 # Atribui-se uma lista para os campos da tabela
                 tab = [
                     doc.get('_id'),
                     doc.get('_rev'),
-                    doc.get('_nome'),
-                    doc.get('_idade'),
-                    doc.get('_cidade')
+                    doc.get('nome'),
+                    doc.get('idade'),
+                    doc.get('cidade')
                 ]
                 # Adiciona-se os campos da tabela a lista de usuários
                 usuarios.append(tab)
                 # Cria-se um cabeçalho da tabela
                 headers = ["ID", "REV", "NOME", "IDADE", "CIDADE"]
-                return print(tabulate(usuarios, headers=headers, tablefmt="simple"))
+            print(tabulate(usuarios, headers=headers, tablefmt="simple"))
     else:
         print('\n>>> Sem conexão com o servidor. <<<')             
 
-# Funcionalidade - inserção de registros
+# Funcionalidade - inserção de novos registros
 def inserir ():
-    
+    print('----------------- NOVO REGISTRO ----------------------')    
     # Atribui-se a conexão
     db = con()
     
@@ -117,75 +118,76 @@ def inserir ():
     else:
         print('\n>>> Não foi possível conectar ao servidor. <<<')
 
-# Funcionalidade - consulta unitária
+# Função assistente - consulta unitária
 def cons_doc (doc_id):
     
-    # Recebe como parâmetro o id fornecido pelo usuário (doc_id)
     # Atribui-se a conexão
     db = con()
     
     # Atribui-se a variável o id
     doc = db.get(doc_id)
     
-    # Conferindo a existência do id no banco
-    if doc:
-        row_doc = [[
+    # Resgatando os dos campos em uma lista 
+    row_doc = [[
             doc.get('_id'),
             doc.get('_rev'),
             doc.get('nome'),
             doc.get('idade'),
             doc.get('cidade')
-        ]]
+    ]]
         # Cria-se um Lista com cabeçalho dos campos
-        headers = ['ID', 'VER', 'NOME', 'IDADE', 'CIDADE']
-        print(tabulate(row_doc, headers=headers, tablefmt='simple'))
-    else:
-        print('\n>>> Falha ao visulizar o registro <<<')        
+    headers = ['ID', 'VER', 'NOME', 'IDADE', 'CIDADE']
+    print(tabulate(row_doc, headers=headers, tablefmt='simple'))
                             
-# Funcionalidade - Atualização de registros                            
+# Funcionalidade - atualizar de registros                            
 def atualizar ():
-    
+    print('------------------------ ATUALIZAR REGISTRO ------------------------------')    
     # Atribui-se a conexão
     db = con()
     
     # Se existir conexão, então...
     if db:
+        # Atribuição recebe entrada do usuário
         doc_id = input('Informe o ID do usuário: ')
-        # chama função de consulta ao usuário
-        cons_doc(doc_id)
-        doc = db.get(doc_id)
         
-        # Tratamento do registro
-        try:
-            if input('>>> Atualizar esse registro (s- sim / n- não)?: ').lower() != 's':
-                print('\n>>> Atualização cancelada. <<<')                
-            else:
-                # Recebe novo nome oou mantém o existente
-                up_nome = input('Atualizar nome: ') or doc['nome']
-                
-                # Recebe nova idade ou mantém a idade existente
-                t_idade = input('Atualizar idade: ')
-                up_idade = int(t_idade) if t_idade else doc['idade']
-                
-                # Recebe nova cidade ou mantém a cidade existente
-                up_cidade = input('Atualizar cidade: ') or doc['cidade']
+        # Testa se o id existir
+        if doc_id in db:
+            # chama função de consulta ao usuário
+            cons_doc(doc_id)
+            doc = db.get(doc_id)
             
-                # Atribuindo a chave
-                doc['nome'] = up_nome
-                doc['idade'] = up_idade
-                doc['cidade'] = up_cidade
+            # Tratamento do registro
+            try:
+                if input('>>> Atualizar esse registro (s- sim / n- não)?: ').lower() != 's':
+                    print('\n>>> Atualização cancelada. <<<')                
+                else:
+                    # Recebe novo nome oou mantém o existente
+                    up_nome = input('Atualizar nome: ') or doc['nome']
+                    
+                    # Recebe nova idade ou mantém a idade existente
+                    t_idade = input('Atualizar idade: ')
+                    up_idade = int(t_idade) if t_idade else doc['idade']
+                    
+                    # Recebe nova cidade ou mantém a cidade existente
+                    up_cidade = input('Atualizar cidade: ') or doc['cidade']
                 
-                # Atualizando no banco
-                db[doc_id] = doc
-                print(f'\n>>> O usuário ({doc[doc_id]}) - {up_nome} foi atualizado com sucesso. <<<')
-        except couchdb.http.ResourceNotFound as e:
-            print(f'Registro não encontrado. {e}')
-    else:
-        print('\n>>> Não foi possível conctar ao servidor. <<<')                
+                    # Atribuindo a chave
+                    doc['nome'] = up_nome
+                    doc['idade'] = up_idade
+                    doc['cidade'] = up_cidade
+                    
+                    # Atualizando no banco
+                    db[doc_id] = doc
+                    print(f'\n>>> O registo: {doc_id} - {up_nome}) foi atualizado com sucesso. <<<')
+            # Exceção caso não encontre registros        
+            except couchdb.http.ResourceNotFound as e:
+                print(f'Registro não encontrado. {e}')
+        else:
+            print('\n>>> Registro inexistente <<<')        
                 
 # Funcionalidade - excluir registro
 def excluir ():
-    
+    print('------------------------ EXCLUIR REGISTRO ------------------------------')    
     # Atribui-se a conexão
     db = con()
     
@@ -193,22 +195,67 @@ def excluir ():
     if db:
         # Recebe ID pelo usuário
         doc_id = input('Informe o ID do usuário: ')
+        
         # chama função de consulta ao usuário
-        cons_doc(doc_id)
-        doc = db.get(doc_id)
+        if doc_id in db:
+            cons_doc(doc_id)
+            doc = db.get(doc_id)
         
-        # Tratamento do registro
+            # Tratamento do registro
+            try:
+                if input('>>> Deseja excuir esse registro: (s- sim / n- não)?: ').lower() != 's':
+                    print('\n>>> Exclusão cancelada. <<<')                
+                else:
+                    db.delete(doc)
+                    print('\n>>> Registro excluído com sucesso! <<<')
+            except couchdb.http.ResourceNotFound as e:
+                print(f'\n>>> Erro ao deletar o arquivo: {e} <<<')
+        else:
+            print('\n>>> Registro inexistente <<<')     
+        
+# Funcionalidade - Painel com menu de opções
+def menu ():
+    
+    print('------------------ SISTEMA DE REGISTROS ------------------')
+    print('1 - Listar usuários')
+    print('2 - Novo usuário')
+    print('3 - Atualizar usuário')
+    print('4 - Excluir usuário')
+    print('5 - Sair')
+    
+    # Loop de nova consulta    
+    while True:
         try:
-            if input('>>> Deseja excuir esse registro: (s- sim / n- não)?: ').lower() != 's':
-                print('\n>>> Exclusão cancelada. <<<')                
+            oc = [1, 2, 3, 4, 5]
+            op = int(input('Opção >>>: '))
+            
+            if op in oc:
+                if op == oc[0]: listar()
+                elif op == oc[1]: inserir()
+                elif op == oc[2]: atualizar()
+                elif op == oc[3]: excluir()
+                elif op == oc[4]:
+                    print('\n>>> Sessão encerrada. <<<')
+                    break
             else:
-                db.delete(doc)
-                print('\n>>> Registro excluído com sucesso! <<<')
-        except couchdb.http.ResourceNotFound as e:
-            print(f'\n>>> Erro ao deletar o arquivo: {e} <<<')
-    else:
-        print('\n>>> Não foi possível conectar ao servidor <<<')     
-        
+                print('\n>>> Opção inválida. Tente novamente: ')
+                continue
+            # Loop interno para tratamento de exceções
+            while True:
+                try:
+                    cons = input('\n>>> Nova consulta (s - sim / n - não)?: ').lower()
+                    if cons == 's':
+                        break
+                    elif cons == 'n': 
+                        print('\n>>> Sessão encerrada. <<<')
+                        return    
+                    else:
+                        raise ValueError ('!!! Opção inválida !!!')
+                except Exception as e:
+                    print(e)    
+        except Exception:
+            print('Entrada inválida: Informe uma das opções (1 a 5)')                    
+
